+++ src/App.css
@@ -1,9 +1,10 @@
+@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
+
 :root {
     --color-1: #231942;
-    --color-2: #5E548E;
-    --color-3: #9F86C0;
-    --color-4: #BE95C4;
-    --color-5: #E0B1CB;
+    --color-2: #9F86C0;
+    --color-3: #E0B1CB;
+    --color-4: #FBFAFF;
 
     --shadow: rgba(0,0,0,0.5);
 
@@ -14,19 +15,31 @@
     --font-1: 12px;
     --font-2: 16px;
     --font-3: 24px;
-
-    --max-width: 300px;
 }
 
 * {
     margin: 0;
     padding: 0;
     box-sizing: border-box;
+
+    font-family: 'Montserrat', sans-serif;
 }
 
 .App {
+    width: 100%;
+    min-height: 100vh;
+
     text-align: center;
+
+    background-image: linear-gradient(to bottom, var(--color-1), var(--color-3));
 }
+
 .Main{
-    margin-top: 80px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    
+    z-index: 1;
+    padding-top: 80px;
 }
