+++ src/components/Cart/Cart.css
@@ -17,7 +17,7 @@
     margin-top: 64px;
     padding-bottom: 128px;
   
-    background-color: var(--color-5);
+    background-color: var(--color-1);
   
     transition: all 0.4s;
 }
