+++ src/components/Filter/JobFilter.css
@@ -1,0 +1,42 @@
+.ProductFilter {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 8px;
+
+    max-width: 320px;
+    height: 24px;
+    margin: 8px 0;
+}
+
+.ProductFilter > select {
+    width: fit-content;
+    height: 100%;
+
+    padding: 0 4px;
+
+    font-weight: bold;
+
+    color: var(--color-1);
+
+    border: none;
+    border-radius: var(--radius);
+
+    outline: none;
+}
+
+.ProductFilter > input {
+    width: 100%;
+    height: 100%;
+
+    padding: 0 4px;
+
+    font-weight: bold;
+
+    color: var(--color-1);
+
+    border: none;
+    border-radius: var(--radius);
+
+    outline: none;
+}
