+++ src/components/Filter/JobFilter.jsx
@@ -1,42 +1,24 @@
 import React from "react";
 import './JobFilter.css';
 
-/*React Icons*/
-import { ImFilter } from 'react-icons/im';
-import { TbArrowsSort } from 'react-icons/tb';
-
-export const JobFilter = ({ 
-        setSelectedBrand, selectedBrand,
-        minimo, setMinimo,
-        maximo, setMaximo
-    }) => {
-    // Update seletedBrand state
+export const JobFilter = ({setSelectedBrand, selectedBrand, minimo, setMinimo, maximo, setMaximo}) => {
     const handleBrandChange = (event) => {
         setSelectedBrand(event.target.value);
     };
+
     return (
         <div className='ProductFilter'>
-            <input type="checkbox" id="ProductFilter-trigger" />
-            <ImFilter />
+            <select
+                value={selectedBrand}
+                onChange={handleBrandChange}>
+                <option value="MENOR">Menor Preço</option>
+                <option value="MAIOR">Maior Preço</option>
+                <option value="CRESCENTE">A - Z</option>
+                <option value="DECRESCENTE">Z - A</option>
+            </select>
 
-            <div className='ProductFilter-container'>
-                <h3>FILTROS</h3>
-
-                <select
-                    value={selectedBrand}
-                    onChange={handleBrandChange}>
-                    <option value="MENOR">Menor Preço</option>
-                    <option value="MAIOR">Maior Preço</option>
-                    <option value="CRESCENTE">A - Z</option>
-                    <option value="DECRESCENTE">Z - A</option>
-                </select>
-
-                <TbArrowsSort />
-
-                <input type="number" placeholder='Valor mínimo' value={minimo} onChange={(e) => setMinimo(e.target.value)} />
-                <input type="number" placeholder='Valor máximo' value={maximo} onChange={(e) => setMaximo(e.target.value)} />
-
-            </div>
+            <input type="number" placeholder='Valor mínimo' value={minimo} onChange={(e) => setMinimo(e.target.value)} />
+            <input type="number" placeholder='Valor máximo' value={maximo} onChange={(e) => setMaximo(e.target.value)} />
         </div>
     )
 }
