--- src/components/Form/FormJob.css
@@ -1,70 +1,0 @@
-form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-evenly;
-    border: solid var(--color-1);
-    border-radius: var(--radius);
-    background-color: var(--color-3);
-    height: 39em;
-    width: 40%;
-}
-form > label{
-    font-weight: bold;
-    font-size: var(--font-3);
-}
-
-form > input {
-    border-radius: var(--radius);
-
-    width: 100%;
-    max-width: var(--max-width);
-
-    padding: var(--padding);
-
-    font-size: var(--font-2);
-}
-
-form>select {
-    
-    outline: none;
-    overflow: hidden;
-
-    font-size: var(--font-2);
-
-    width: 100%;
-    max-width: var(--max-width);
-
-    border-radius: var(--radius);
-    background-color: var(--color-5);
-}
-
-form>option {
-    width: 100%;
-
-    padding: var(--padding);
-
-    text-align: center;
-
-    
-    
-}
-
-form > button {
-    height: 3rem;
-
-    padding: var(--padding);
-
-    font-size: var(--font-3);
-
-    font-weight: bold;
-
-    border-radius: var(--radius);
-
-    
-}
-form>button:hover{
-    background-color: var(--color-1);
-}
-
-
