+++ src/components/Header/HeaderBox.css
@@ -15,32 +15,32 @@ header {
   padding-left: 16px;
  
   background-color: var(--color-1);
-  
-  box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.5);
 
   user-select: none;
-  margin-top: -80px;
+
+  box-shadow: 0px 0px 4px var(--shadow);
 }
   
 header > .header-logo {
   cursor: pointer;
-  font-size: 32px;
-  color: var(--color-5);
+  font-size: 28px;
+  color: var(--color-4);
 }
   
 header > nav {
   display: flex;
   align-items: center;
   justify-content: center;
-  gap: 32px;
- 
-  color: var(--color-5);
+  gap: 16px;
 }
   
 header > nav > span {
   cursor: pointer;
-  font-size: 20px;
+
+  font-size: 16px;
   font-weight: bold;
+
+  color: var(--color-4);
 }
 
 header > nav > .cart-icon-counter {
@@ -57,9 +57,10 @@ header > nav > .cart-icon-counter {
 
   font-size: 32px;
   font-weight: bold;
-  color: var(--color-1);
 
-  background-color: var(--color-5);
+  color: var(--color-4);
+
+  background-color: var(--color-2);
 
   border-radius: var(--radius) 0 0 var(--radius);
 
@@ -67,17 +68,25 @@ header > nav > .cart-icon-counter {
 }
 
 header > nav > .cart-icon-counter > span {
-  font-size: 24px;
-  font-weight: bold;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: fit-content;
+  height: 100%;
+
+  margin: 0;
+  padding: 0;
+  font-size: 20px;
 }
   
-header > nav > .cart-icon {
-  font-size: 32px;
+header > nav > .cart-icon-counter > .cart-icon {
+  font-size: 28px;
 
   transition: all 0.4s;
 }
 
-header > nav > .cart-icon:active {
+header > nav > .cart-icon-counter > .cart-icon:active {
   transform: scale(0.9);
 }
 
@@ -92,7 +101,7 @@ header > nav > .cart-icon:active {
   position: absolute;
   right: 0px;
 
-  width: 96px;
+  width: 80px;
   height: 64px;
 
   opacity: 0;
@@ -100,4 +109,8 @@ header > nav > .cart-icon:active {
 
 .trigger-cart-modal:checked ~ .cart-modal {
   left: 0;
+}
+
+@media (max-width: 480px) {
+
 }
