+++ src/components/Header/HeaderBox.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect } from "react";
 import './HeaderBox.css';
 
 /*Icons*/
@@ -12,21 +12,31 @@ export const HeaderBox = ({quantItensCart, setQuantItensCart, cartList, setCartL
 
     useEffect(() => {
         let icon = document.getElementById('cart-box')
+        let checkbox = document.getElementById('trigger-cart')
 
         if (quantItensCart === 0) {
-            icon.style.marginRight = '-24px';
+            icon.style.marginRight = '-24px'
+            checkbox.style.marginRight = '-24px'
         } else {
-            icon.style.marginRight = '0';
+            icon.style.marginRight = '0'
+            checkbox.style.marginRight = '0'
         }
     })
+
+    const closeCart = (pageGoTo) => {
+        document.getElementById('trigger-cart').checked = false
+
+        setPage(pageGoTo)
+    }
 
     return (
         <header>
             <SiCodingninjas className="header-logo"/>
 
             <nav>
-                <span onClick={() => setPage("how")}>Como funciona?</span>
-                <span onClick={() => setPage("services")}>Serviços</span>
+                <span onClick={() => closeCart('form')}>Cadastro</span>
+                <span onClick={() => closeCart('how')}>Como funciona?</span>
+                <span onClick={() => closeCart('services')}>Serviços</span>
 
                 <div className="cart-icon-counter" id="cart-box">
                     <HiOutlineShoppingCart className="cart-icon"/>
@@ -34,11 +44,11 @@ export const HeaderBox = ({quantItensCart, setQuantItensCart, cartList, setCartL
                 </div>  
             </nav>
 
-            <input type="checkbox" className="trigger-cart-modal"/>
+            <input type="checkbox" className="trigger-cart-modal" id="trigger-cart"/>
             <CartModal
                 className="cart-modal"
                 quantItensCart={quantItensCart} 
-                setQuantItensCart={setQuantItensCart} 
+                setQuantItensCart={setQuantItensCart}
 
                 cartList={cartList}
                 setCartList={setCartList}
