+++ src/components/Search/SearchInput.css
@@ -1,0 +1,53 @@
+.search-input {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    max-width: 320px;
+    width: 100%;
+    height: 32px;
+}
+
+.search-input > input {
+    width: 100%;
+    height: 100%;
+
+    font-size: var(--font-2);
+    font-weight: bold;
+
+    text-align: center;
+
+    color: var(--color-1);
+
+    outline: none;
+
+    border: none;
+    border-radius: var(--radius);
+}
+
+.search-input > input::-webkit-search-cancel-button {
+    display: none;
+}
+
+.search-input > .search-icon {    
+    position: absolute;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    height: fit-content;
+
+    margin-left: 140px;    
+
+    font-size: 20px;
+
+    color: var(--color-1);
+}
+
+@media (max-width: 320px) {
+    .search-input > .search-icon {
+        margin-left: 0;
+        right: 16px;
+    }
+}
